moved files in ProfilePage directory out
